requirements.txt
Django>=4.2,<5.0
django-ckeditor>=6.7.0
django-crispy-forms>=2.0
django-js-asset>=2.1.0
pillow>=9.0.0
